Complete rewrite of Javascript rules.
New scheme bundles Webpack into a single file, which is then used for building more rules.
That solves a bunch of correctness issues about searching through plz-out/gen for node modules,
and will in future make it easier to have multiple invocations of it.

Yarn rules are rewritten also, and are much more akin to maven_jar or python_wheel now.

Still quite a lot of TODOs etc...